Rename all databased setting classes to be specific to rulesets for now
